ai-voice/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
fastapi==0.124.4
filelock==3.20.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
omegaconf==2.3.0
openai-whisper==20250625
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
scipy==1.16.3
silero==0.5.2
soundfile==0.13.1
starlette==0.50.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
